backend - Copy/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pymongo>=4.6.0
passlib>=1.7.4
python-jose>=3.3.0
google-genai>=0.3.0
requests>=2.31.0
email-validator>=2.0.0
python-multipart>=0.0.6
